% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{call_drug_hits}
\alias{call_drug_hits}
\title{Call significant responses for scored chemogenomic data.}
\usage{
call_drug_hits(
  scores,
  control_screen_name = NULL,
  condition_screen_names = NULL,
  fdr_threshold = 0.1,
  differential_threshold = 0,
  neg_type = "Negative",
  pos_type = "Positive"
)
}
\arguments{
\item{scores}{Dataframe returned from \code{score_drugs_vs_control} or 
\code{score_drugs_vs_controls}.}

\item{control_screen_name}{Name of a control screen to test condition screens against, 
or NULL to score data returned from \code{score_drugs_vs_controls}.}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}
}
\value{
Dataframe of scored data with differential effects called as significant
  for the specified conditions.
}
\description{
Run this to call significant responses for data returned from 
\code{score_drugs_vs_control} or \code{score_drugs_vs_controls}.
}
