% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_group_file}
\alias{check_group_file}
\title{Checks a group file}
\usage{
check_group_file(group_file, screens)
}
\arguments{
\item{group_file}{Path to .tsv file mapping screens to their controls for scoring, with four 
columns for "Screen", "Control", "Group" and "Type." The "Group" column can contain any 
meaningful label for each group, whereas the "Type" column must be labeled either "control" 
or "condition." For each unique group label, all conditions in that group are scored
against all controls.}
}
\value{
TRUE.
}
\description{
Checks to make sure that the contents of the .tsv file and its formats are appropriate for
running group scoring functions.
}
\keyword{internal}
