% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qGI_utils.R
\name{compute_control_weights}
\alias{compute_control_weights}
\title{Computes weights for control screens.}
\usage{
compute_control_weights(
  control_df,
  condition_df,
  control_names,
  condition_names,
  matched_controls,
  method = "exp",
  matched_fraction = 0.75
)
}
\arguments{
\item{control_df}{LFC dataframe for control screens.}

\item{condition_df}{LFC dataframe for condition screens.}

\item{control_names}{List of screen names for control screens to compute effect
sizes for.}

\item{condition_names}{List of screen names for control screens to compute effect
sizes for.}

\item{matched_controls}{A vector of control screen names corresponding to all 
condition names, which must be the same length as condition_names.}

\item{method}{Either "linear" to give equal weight to all other control screens
or "exp" to give exponentially-decreasing weight to all other control screens,
based on those with the most similar LFCs (default "linear").}

\item{matched_fraction}{A value between 0 and 1 specifying the weight to give
the matched control for each condition, where weights for all other control
screens will sum to 1 - matched_fraction (default 0.75).}
}
\value{
Dataframe of weights for each condition and control with the number of 
  rows equal to length(condition_names) and the number of columns equal to 
  length(control_names).
}
\description{
Computes weights for control screens based on matching condition screens
for \code{score_drugs_vs_controls}.
}
