% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{pca_screens}
\alias{pca_screens}
\title{PCA-normalizes residual LFCs for the given screens}
\usage{
pca_screens(
  df,
  cols,
  n_components,
  scale = FALSE,
  na_behavior = "mean_replace",
  exclude_screens = NULL
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{cols}{Numerical column names to normalize with PCA.}

\item{n_components}{Vector containing indices of principal components to remove from data.}

\item{scale}{Whether or not to scale replicates before extracting principal 
components (default FALSE).}

\item{na_behavior}{Whether to replace NAs with row (per-guide) mean values or to
omit NAs, as either "mean_replace" or "omit" (default "mean_replace").}

\item{exclude_screens}{A list of screen names to exclude, e.g. for replicates with
mostly NA values (default NULL).}
}
\value{
PCA-normalized dataframe.
}
\description{
PCA-normalizes residual LFCs for a dataset by extracting a given number of principal
components from the dataset, projecting the dataset to those components, and 
subtracting the projected dataset from the original dataset. After performing 
PCA-normalization, consider re-centering LFCs by the mean of non-essential genes.
}
