% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_chromosome}
\alias{plot_chromosome}
\title{Plots chromosomal QC.}
\usage{
plot_chromosome(df, guide_df, output_folder, plot_type = "png")
}
\arguments{
\item{df}{LFC dataframe passed into \code{score_drugs_vs_controls} with chromosomal location 
specified in a column named "chr", start coordinates specified in "start_loc" and stop 
coordinates specified in "stop_loc."}

\item{guide_df}{Dataframe of guide-level qGI scores.}

\item{output_folder}{Folder to output plots to.}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Plots qGI effects by chromosomal location.
}
