% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_drug_response}
\alias{plot_drug_response}
\title{Plots drug response for scored data.}
\usage{
plot_drug_response(
  scores,
  control_name = NULL,
  condition_name = NULL,
  output_folder = NULL,
  loess = TRUE,
  neg_type = "Negative",
  pos_type = "Positive",
  label_fdr_threshold = NULL,
  plot_type = "png"
)
}
\arguments{
\item{scores}{Dataframe of scores returned from \code{call_drug_hits}.}

\item{control_name}{Name of control passed to \code{call_drug_hits}, or 
NULL for data scored by \code{score_drugs_vs_controls} (default NULL).}

\item{condition_name}{Name of condition passed to \code{call_drug_hits}.}

\item{output_folder}{Folder to output plots to.}

\item{loess}{If true and data was loess-normalized, plots loess null model instead
(default TRUE).}

\item{neg_type}{Label for significant effects with a negative differential effect
passed to \code{call_drug_hits} (default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
passed to \code{call_drug_hits} (default "Positive").}

\item{label_fdr_threshold}{Threshold below which to plot gene labels for significant
hits, or NULL to plot without labels (default NULL).}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}
}
\description{
Pretty-plots response for chemogenomic (e.g. for directly comparing drug 
response to DMSO response). Assumes that data was scored by 
\code{score_drugs_vs_control} and significant effects were called by 
\code{call_drug_hits}.
}
