% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scree}
\alias{plot_scree}
\title{Makes scree plot}
\usage{
plot_scree(
  df,
  cols,
  scale = FALSE,
  na_behavior = "mean_replace",
  exclude_screens = NULL
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{cols}{Numerical column names to normalize with PCA.}

\item{scale}{Whether or not to scale replicates before extracting principal 
components (default FALSE).}

\item{na_behavior}{Whether to replace NAs with row (per-guide) mean values or to
omit NAs, as either "mean_replace" or "omit" (default "mean_replace").}

\item{exclude_screens}{A list of screen names to exclude, e.g. for replicates with
mostly NA values (default NULL).}
}
\value{
Scree plot as a ggplot2 object.
}
\description{
Outputs scree plot for a dataset to help select a number of principal components to
remove.
}
