% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_drugs_batch}
\alias{score_drugs_batch}
\title{Scores multiple drugs against multiple controls}
\usage{
score_drugs_batch(
  df,
  screens,
  batch_file,
  output_folder,
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  ma_transform = TRUE,
  control_genes = c("None", ""),
  n_components = 0,
  chromosomal_correction = FALSE,
  weight_method = "exp",
  matched_fraction = 0.75,
  sd_scale_factor = NULL,
  fdr_method = "BY",
  fdr_threshold = 0.1,
  differential_threshold = 0.5,
  neg_type = "Negative",
  pos_type = "Positive",
  label_fdr_threshold = NULL,
  save_residuals = FALSE,
  plot_residuals = TRUE,
  plot_type = "png",
  verbose = FALSE
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{batch_file}{Path to .tsv file mapping screens to their controls for scoring, with two 
columns for "Screen" and "Control." Screens to score against dervied null-models with the
combn scoring mode must have their respective control labeled as "combn."}

\item{output_folder}{Folder to output scored data and plots to.}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{n_components}{Vector containing indices of principal components to remove from data.}

\item{chromosomal_correction}{If TRUE, corrects chromosomal shifts by down-weighting qGI
scores for shifted regions (default FALSE).}

\item{weight_method}{If "linear" weights non-matched control screens equally. If exponential,
instead weights non-matched control screens according to a decreasing exponential function
based on the similarity of control screens to the given condition screen.}

\item{matched_fraction}{The weight given to the matched control as a fraction of 1, where
all non-matched controls receive a total weight equal to 1 - matched_fraction (default 0.75).}

\item{sd_scale_factor}{Factor to normalize SDs against for scaling. If NULL for group scoring,
the mean is computed across guide-level residuals, otherwise the given scalar is used instead.
If NULL for one-off scoring, this operation is not performed (default NULL).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default
"BY")}

\item{fdr_threshold}{Threshold below which to call gene effects as significant 
(default 0.1).}

\item{differential_threshold}{Absolute value threshold on differential effects, 
below which gene effects are not called as significant (default 0.5).}

\item{neg_type}{Label for significant effects with a negative differential effect
(default "Negative").}

\item{pos_type}{Label for significant effects with a positive differential effect
(default "Positive").}

\item{label_fdr_threshold}{Threshold below which to plot gene labels for significant
hits, or NULL to plot without labels (default NULL).}

\item{save_residuals}{If true, saves residuals for each screen to the output folder
(default FALSE).}

\item{plot_residuals}{If true, plots residual effects for all top hits (default TRUE).}

\item{plot_type}{Type of plot to output, one of "png" or "pdf" (default "png").}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\description{
Takes in an input .tsv file with two columns for "Screen" and "Control" and scores
all screens listed in "Screen" against their corresponding screens listed in
"Control." Outputs all files and plots in the specified folder.
}
