% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_drugs_vs_control}
\alias{score_drugs_vs_control}
\title{Scores conditions against a single control.}
\usage{
score_drugs_vs_control(
  df,
  screens,
  control_screen_name,
  condition_screen_names,
  control_genes = c("None", ""),
  min_guides = 3,
  test = "moderated-t",
  loess = TRUE,
  ma_transform = TRUE,
  fdr_method = "BY",
  sd_scale_factor = NULL,
  return_residuals = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{LFC dataframe.}

\item{screens}{List of screens generated with \code{add_screens}.}

\item{control_screen_name}{Name of a control screen to test condition screens against.}

\item{condition_screen_names}{A list of condition screen names to score against the 
control screen.}

\item{control_genes}{List of control genes to remove, e.g. "luciferase" (default c("None", "")).}

\item{min_guides}{The minimum number of guides per gene pair required to score data 
(default 3).}

\item{test}{Type of hypothesis testing to run. Must be one of "rank-sum" for Wilcoxon
rank-sum testing or "moderated-t" for moderated t-testing (default "moderated-t").}

\item{loess}{If true, loess-normalizes residuals before running hypothesis testing.
Only works when test = "moderated-t" (default TRUE).}

\item{ma_transform}{If true, M-A transforms data before running loess normalization. Only
has an effect when loess = TRUE (default TRUE).}

\item{fdr_method}{Type of FDR to compute. One of "BH", "BY" or "bonferroni" (default "BY").}

\item{sd_scale_factor}{Factor to normalize SDs against for scaling. If NULL, this operation
is not performed - this behavior is different for group scoring! (default NULL).}

\item{return_residuals}{If FALSE, returns NA instead of residuals dataframe (default TRUE).
This is recommend if scoring large datasets and memory is a limitation.}

\item{verbose}{If true, prints verbose output (default FALSE).}
}
\value{
A list containing two dataframes. The first entry, named "scored_data" in the list,
  contains scored data with separate columns given by the specified control and condition
  names. The second entry, named "residuals" in the list, is a dataframe containing control,
  condition and loess-normalized residuals for all guides.
}
\description{
Scores guides for any number of drug screens against a control screen
(e.g. for directly comparing drug response to DMSO response). After running 
this function, pass the resulting dataframe to \code{call_drug_hits} to 
call significant effects.
}
